The commit contains these commits from master:
* 52cd1d7 Add gtest-1.7.0 as submodule
* 3b136a0 Change function name to specify udp logic.
* 6c49673 linux: Avoid empty initializers brances
* e33a183 linux: Fix mistake use OC_REP_BOOL not BOOL
* 00d8bae Apply CoAP over TCP logic.
* 0840e0a tests: use __func__ for gcc-7.2
* b186c73 ci: Adjust relative paths for makefile wrapper
* 3e617ad linux: Use unused args
* 1927383 ci: Add auto_build scripts
* a8a5b21 Add TCP scheme(coap+tcp) for endpoint
* abecadd Add CoAP over TCP feature.

+++ include/oc_blockwise.h
@@ -97,13 +97,13 @@ void oc_blockwise_free_response_buffer(oc_blockwise_state_t *buffer);
 
 const void *oc_blockwise_dispatch_block(oc_blockwise_state_t *buffer,
                                         uint32_t block_offset,
-                                        uint16_t requested_block_size,
-                                        uint16_t *payload_size);
+                                        uint32_t requested_block_size,
+                                        uint32_t *payload_size);
 
 bool oc_blockwise_handle_block(oc_blockwise_state_t *buffer,
                                uint32_t incoming_block_offset,
                                const uint8_t *incoming_block,
-                               uint16_t incoming_block_size);
+                               uint32_t incoming_block_size);
 
 void oc_blockwise_scrub_buffers();
